cleaned up code to adm.py
